+++ js/get-randon-number.js
@@ -1,0 +1,17 @@
+const getRandomPositiveInteger = (min, max)=> {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomPositiveFloat = (min, max, digits = 1)=> {
+  const lower = Math.min(Math.abs(min), Math.abs(max));
+  const upper = Math.max(Math.abs(min), Math.abs(max));
+  const result = Math.random() * (upper - lower) + lower;
+
+  return result.toFixed(digits);
+};
+
+export {getRandomPositiveFloat, getRandomPositiveInteger};
